agentcore_deploy/requirements.txt
boto3
botocore
opensearch-py==2.7.1
pydantic==2.10.2
ipython==8.29.0
pyyaml
retrying
strands-agents
strands-agents-tools
mcp
bedrock-agentcore
bedrock-agentcore-starter-toolkit
pydantic==2.10.2
pyyaml
retrying
